Fix #2710: reject subnormal query vectors instead of searching zeros (#2715)
A query embedding component below f32 range (|x| < ~1.2e-38) was silently cast
to 0.0 during deserialization, so the search ran against a zeroed vector and
returned wrong results with no error. The query-side counterpart of #2689:
widen VectorQuery/VectorIndexQuery `query` to Vec<f64> and validate via the
engine `from_wire` constructor, rejecting underflow/overflow with
invalid_argument.engine.vector_embedding.

Invariants verified: none apply (transport-layer input validation; accepted
queries execute identically, rejected queries error before the search runs)
Tests added: vector_query_rejects_a_subnormal_query_vector_instead_of_searching_zeros

+++ crates/executor/src/command.rs
@@ -825,8 +825,9 @@ pub enum Command {
         space: Option<String>,
         /// Collection name.
         collection: String,
-        /// Query embedding.
-        query: Vec<f32>,
+        /// Query embedding. Accepted at wire (f64) precision and narrowed to the
+        /// searched f32; a value that underflows or overflows f32 is rejected.
+        query: Vec<f64>,
         /// Maximum number of matches.
         k: u64,
         /// Optional metadata filter.
@@ -846,8 +847,9 @@ pub enum Command {
         space: Option<String>,
         /// Collection name.
         collection: String,
-        /// Query embedding.
-        query: Vec<f32>,
+        /// Query embedding. Accepted at wire (f64) precision and narrowed to the
+        /// searched f32; a value that underflows or overflows f32 is rejected.
+        query: Vec<f64>,
         /// Maximum number of matches.
         k: u64,
         /// Optional metadata filter.
